+++ src/__screenshot_tests__/__image_snapshots__/list-screenshot-test-tsx-ordered-list-1-snap.png
