+++ src/__screenshot_tests__/__image_snapshots__/list-screenshot-test-tsx-ordered-list-default-icons-2-snap.png
