+++ src/__screenshot_tests__/__image_snapshots__/list-screenshot-test-tsx-unordered-list-without-marker-1-snap.png
