+++ src/index.tsx
@@ -53,8 +53,7 @@ export {default as Tag} from './tag';
 export type {TagType} from './tag';
 export {Placeholder} from './placeholder';
 export {Title1, Title2, Title3, Title4} from './title';
-export {RowList, Row, BoxedRowList, BoxedRow} from './list';
-export {Ul, Ol, Li} from './ul';
+export {RowList, Row, BoxedRowList, BoxedRow, UnorderedList, OrderedList, ListItem} from './list';
 export {default as Switch} from './switch-component';
 export {default as Checkbox} from './checkbox';
 export {default as RadioButton, RadioGroup} from './radio-button';
