+++ src/list.css.ts
@@ -192,3 +192,99 @@ export const dualActionRight = style([
         lineHeight: 0,
     },
 ]);
+
+const ulVerticalGap = 8;
+const ulVerticalGapDesktop = 16;
+
+export const ul = style({
+    display: 'block',
+    margin: 0,
+    padding: 0,
+
+    selectors: {
+        'li &': {
+            paddingTop: ulVerticalGap,
+            '@media': {
+                [mq.desktopOrBigger]: {
+                    paddingTop: ulVerticalGapDesktop,
+                },
+            },
+        },
+    },
+});
+
+export const liWithCustomIcon = style({
+    display: 'flex',
+    alignItems: 'flex-start',
+    paddingLeft: 8,
+    paddingBottom: ulVerticalGap,
+    '@media': {
+        [mq.desktopOrBigger]: {
+            paddingLeft: 16,
+            paddingBottom: ulVerticalGapDesktop,
+        },
+    },
+    ':last-child': {
+        paddingBottom: 0,
+    },
+});
+
+export const liWithoutMarker = style({
+    display: 'flex',
+    alignItems: 'flex-start',
+    paddingLeft: 'calc(16px + 1em)',
+    paddingBottom: ulVerticalGap,
+    '@media': {
+        [mq.desktopOrBigger]: {
+            paddingLeft: 'calc(32px + 1em)',
+            paddingBottom: ulVerticalGapDesktop,
+        },
+    },
+    ':last-child': {
+        paddingBottom: 0,
+    },
+});
+
+export const li = style({
+    marginLeft: 24,
+    paddingBottom: ulVerticalGap,
+    '@media': {
+        [mq.desktopOrBigger]: {
+            marginLeft: 32,
+            paddingLeft: 8,
+            paddingBottom: ulVerticalGapDesktop,
+        },
+    },
+    ':last-child': {
+        paddingBottom: 0,
+    },
+
+    selectors: {
+        'ul > &': {
+            listStyleType: 'disc',
+        },
+        'ol > &': {
+            listStyleType: 'decimal',
+        },
+        'ul ul > &': {
+            listStyleType: 'circle',
+        },
+        'ul ul ul > &': {
+            listStyleType: 'square',
+        },
+        'ol ol > &': {
+            listStyleType: 'lower-alpha',
+        },
+        'ol ol ol > &': {
+            listStyleType: 'lower-roman',
+        },
+    },
+});
+
+export const liContent = style({
+    selectors: {
+        [`${liWithCustomIcon} &`]: {
+            flex: 1,
+        },
+    },
+});
